refactor: move getMcpPermissionToolName to types; drop rawInput fallback
Project policy (CLAUDE.md) bans reading from Claude Code SDK's rawInput
for metadata. The handler now writes _meta.claudeCode.toolName on every
MCP approval request, so the rawInput compatibility fallback is no
longer needed.

Generated-By: PostHog Code
Task-Id: c5674c04-c95c-4bfe-bfcf-248a921b4aae
